backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
firebase-admin==6.6.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0
requests==2.32.3
anthropic==0.52.0
openai==1.82.0
gunicorn==23.0.0
